fix: manual parsing or ARP headers instead of using memcpy
to resolve byte-swapping issues in ARP header multi-byte integers

+++ python/hex/protocols.py
@@ -85,7 +85,7 @@ class EtherHeader(ctypes.Structure):
         ('src_mac', 	CT_MAC_ADDRESS), 
         ('dst_mac', 	CT_MAC_ADDRESS),
         ('type', 		ctypes.c_uint16),
-        ('len', 		ctypes.c_size_t)
+        ('len', 		ctypes.c_uint32)
     ]
 
 class IPV4Header(ctypes.Structure):
